refactor(parser): split expr/postfix.rs into cohesive submodules (§7-8)
Split the 3803-line parser/expr/postfix.rs (over the 500-line limit) into a
facade plus five themed submodules under src/parser/expr/postfix/. All but
one are under 500 lines; loop_.rs (2665) holds the single ~2200-line
postfix_expr_loop function, which cannot be split at a function boundary.
postfix.rs is free parser functions addressed as postfix::<name>; the facade
re-exports each submodule preserving its original visibility, so no external
caller changes. expr/mod.rs is untouched.

- mod.rs (20)          - facade: mod decls + re-exports preserving visibility
- helpers.rs (131)     - small utilities (negative literal, angle-key, whatevercode)
- adverb.rs (418)      - subscript adverb parsing (:exists, :delete, :k, :kv, etc.)
- call_method.rs (395) - bracket indices, QuotedMethodName, postfix operator helpers
- dot_assign.rs (243)  - .= mutating method call: atomic_var_name, wrap_dot_assign, parse_dot_assign
- loop_.rs (2665)      - prefix_expr, postfix entry points, postfix_expr_loop (single indivisible ~2200-line fn), parse_superscript_exp

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility. Body refs to expr-level siblings rewritten to
absolute crate::parser paths (submodules are one level deeper).
The fn multiset is byte-identical. No behavior change.
