+++ src/parser/expr/postfix/loop_.rs
@@ -1,1185 +1,34 @@
-use super::super::helpers::{
-    consume_unspace, is_ident_char, is_non_breaking_space, split_angle_words, ws,
+use super::adverb::{
+    DeleteAdverb, build_adverb_error_call, collect_remaining_adverbs, determine_subscript_what,
+    multidim_target_var_name, normalize_adverb_name, parse_delete_adverb,
+    parse_dynamic_subscript_adverb, parse_subscript_adverb_with_expr,
+    subscript_adverb_expr_with_cond, supports_postfix_call_adverbs, try_parse_exists_adverb,
+    try_parse_unknown_adverb,
 };
-use super::super::parse_result::{PError, PResult, parse_char, take_while1};
-use super::super::primary::{colonpair_expr, parse_block_body, parse_call_arg_list, primary};
-
+use super::call_method::{
+    ParsedBracketIndex, QuotedMethodName, append_call_arg, auto_invoke_bareword_method_target,
+    has_ternary_else_after, is_postfix_operator_boundary, parse_bracket_indices,
+    parse_bracket_indices_inner, parse_custom_postfix_operator, parse_prefix_as_postfix,
+    parse_private_method_name, parse_quoted_method_name,
+};
+use super::dot_assign::{atomic_var_name, parse_dot_assign};
+use super::helpers::{
+    compose_prefix_into_whatevercode, extract_negative_literal, extract_range_negative_end,
+    is_angle_key_char, make_negative_subscript_error,
+};
 use crate::ast::{ExistsAdverb, Expr, HyperSliceAdverb, Stmt};
+use crate::parser::expr::operators::{
+    PrefixUnaryOp, parse_postfix_update_op, parse_prefix_unary_op,
+};
+use crate::parser::expr::{expression, expression_no_sequence};
+use crate::parser::helpers::{consume_unspace, is_ident_char, split_angle_words, ws};
+use crate::parser::parse_result::{PError, PResult, parse_char, take_while1};
+use crate::parser::primary::{colonpair_expr, parse_block_body, parse_call_arg_list, primary};
 use crate::symbol::Symbol;
 use crate::token_kind::TokenKind;
 use crate::value::Value;
 
-use super::operators::{PrefixUnaryOp, parse_postfix_update_op, parse_prefix_unary_op};
-use super::{expression, expression_no_sequence};
-
-/// Check if an expression is a negative integer literal (e.g., `-(1)` from `-1`).
-/// Returns the negative value string (e.g., "-1") if so.
-fn extract_negative_literal(expr: &Expr) -> Option<String> {
-    if let Expr::Unary {
-        op: TokenKind::Minus,
-        expr: inner,
-    } = expr
-    {
-        if let Expr::Literal(Value::Int(n)) = inner.as_ref() {
-            return Some(format!("-{}", n));
-        }
-        if let Expr::Literal(Value::BigInt(n)) = inner.as_ref() {
-            return Some(format!("-{}", n));
-        }
-    }
-    None
-}
-
-/// Check if an expression is a range ending in a negative literal (e.g., `0..-1`).
-/// Returns the negative end value string if so.
-fn extract_range_negative_end(expr: &Expr) -> Option<String> {
-    if let Expr::Binary {
-        op: TokenKind::DotDot | TokenKind::DotDotCaret,
-        right,
-        ..
-    } = expr
-    {
-        return extract_negative_literal(right);
-    }
-    None
-}
-
-/// Build a fatal X::Obsolete parse error for negative subscripts.
-fn make_negative_subscript_error(neg_val: &str) -> PError {
-    let old = format!("a negative {} subscript to index from the end", neg_val);
-    let replacement = format!("a function such as *{}", neg_val);
-    let message = format!(
-        "X::Obsolete: Unsupported use of {}. In Raku please use: {}.",
-        old, replacement
-    );
-    let err = crate::value::RuntimeError::obsolete(&old, &replacement);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("old".to_string(), Value::str(old));
-    attrs.insert("replacement".to_string(), Value::str(replacement));
-    attrs.insert("message".to_string(), Value::str(err.message.clone()));
-    let exception = Value::make_instance(crate::symbol::Symbol::intern("X::Obsolete"), attrs);
-    PError::fatal_with_exception(message, Box::new(exception))
-}
-
-/// Check if a character is valid inside an angle-bracket word key (`<key>`).
-/// Raku allows any non-whitespace character that isn't `>` in angle bracket words.
-/// We allow: alphanumeric, common ASCII punctuation used in identifiers/keys,
-/// non-breaking spaces, and any non-ASCII Unicode character that isn't whitespace.
-fn is_angle_key_char(c: char) -> bool {
-    c.is_alphanumeric()
-        || c == '_'
-        || c == '-'
-        || c == '!'
-        || c == '*'
-        || c == '.'
-        || c == ':'
-        || c == '?'
-        || c == '+'
-        || c == '/'
-        || c == '$'
-        || c == '@'
-        || c == '%'
-        || c == '&'
-        || is_non_breaking_space(c)
-        || (!c.is_ascii() && !c.is_whitespace())
-}
-
-/// When a prefix operator is applied to a WhateverCode (Lambda or AnonSubParams),
-/// compose the prefix into the body so that `+(* + 1)` becomes `-> $_ { +($_ + 1) }`
-/// instead of trying to numify the closure itself.
-fn compose_prefix_into_whatevercode(op: TokenKind, expr: Expr) -> Expr {
-    match expr {
-        Expr::Lambda {
-            param,
-            mut body,
-            is_whatever_code,
-        } => {
-            wrap_last_stmt_with_unary(&mut body, op.clone());
-            Expr::Lambda {
-                param,
-                body,
-                is_whatever_code,
-            }
-        }
-        Expr::AnonSubParams {
-            params,
-            param_defs,
-            return_type,
-            mut body,
-            is_rw,
-            is_whatever_code,
-        } => {
-            wrap_last_stmt_with_unary(&mut body, op.clone());
-            Expr::AnonSubParams {
-                params,
-                param_defs,
-                return_type,
-                body,
-                is_rw,
-                is_whatever_code,
-            }
-        }
-        other => Expr::Unary {
-            op,
-            expr: Box::new(other),
-        },
-    }
-}
-
-/// Wrap the last expression statement in a list of statements with a unary operator.
-fn wrap_last_stmt_with_unary(stmts: &mut [Stmt], op: TokenKind) {
-    if let Some(Stmt::Expr(expr)) = stmts.last_mut() {
-        let inner = std::mem::replace(expr, Expr::Literal(Value::Nil));
-        *expr = Expr::Unary {
-            op,
-            expr: Box::new(inner),
-        };
-    }
-}
-
-/// Try to parse a secondary adverb after :exists/:!exists.
-/// Returns (remaining_input, adverb).
-fn parse_exists_secondary_adverb(input: &str) -> (&str, ExistsAdverb) {
-    if input.starts_with(":!kv") && !is_ident_char(input.as_bytes().get(4).copied()) {
-        return (&input[4..], ExistsAdverb::NotKv);
-    }
-    if input.starts_with(":kv") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return (&input[3..], ExistsAdverb::Kv);
-    }
-    if input.starts_with(":!p") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return (&input[3..], ExistsAdverb::NotP);
-    }
-    if input.starts_with(":p") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return (&input[2..], ExistsAdverb::P);
-    }
-    if input.starts_with(":!v") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return (&input[3..], ExistsAdverb::NotV);
-    }
-    if input.starts_with(":!k") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return (&input[3..], ExistsAdverb::InvalidNotK);
-    }
-    if input.starts_with(":k") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return (&input[2..], ExistsAdverb::InvalidK);
-    }
-    if input.starts_with(":v") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return (&input[2..], ExistsAdverb::InvalidV);
-    }
-    (input, ExistsAdverb::None)
-}
-
-/// Parse dynamic subscript adverb: :$delete, :$exists — the variable value
-/// determines whether the adverb is active at runtime.
-fn parse_dynamic_subscript_adverb(input: &str) -> Option<&str> {
-    if !input.starts_with(":$") {
-        return None;
-    }
-    let r = &input[2..];
-    // Read identifier
-    let end = r
-        .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-        .unwrap_or(r.len());
-    if end == 0 {
-        return None;
-    }
-    let name = &r[..end];
-    // Only recognize known subscript adverb names
-    match name {
-        "delete" | "exists" => Some(&r[end..]),
-        _ => None,
-    }
-}
-
-/// Parse a subscript adverb like `:k`, `:!kv`, `:k($ok)`, `:kv(0)`, etc.
-/// Returns (remaining_input, mode_string, optional_dynamic_expr).
-/// When the adverb has a dynamic expression argument (e.g. `:k($var)`),
-/// the third element is `Some(expr)` and the mode is the base adverb name.
-fn parse_subscript_adverb_with_expr(input: &str) -> Option<(&str, &'static str, Option<Expr>)> {
-    if input.starts_with(":!kv") && !is_ident_char(input.as_bytes().get(4).copied()) {
-        return Some((&input[4..], "not-kv", None));
-    }
-    if input.starts_with(":!p") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return Some((&input[3..], "not-p", None));
-    }
-    if input.starts_with(":!k") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return Some((&input[3..], "not-k", None));
-    }
-    if input.starts_with(":!v") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return Some((&input[3..], "not-v", None));
-    }
-    if let Some(rest) = input.strip_prefix(":kv(0)") {
-        return Some((rest, "kv0", None));
-    }
-    if let Some(rest) = input.strip_prefix(":kv(1)") {
-        return Some((rest, "kv", None));
-    }
-    // :kv(expr) — dynamic adverb with expression
-    if input.starts_with(":kv(")
-        && let Some((rest, expr)) = try_parse_adverb_expr(&input[3..])
-    {
-        return Some((rest, "kv", Some(expr)));
-    }
-    if input.starts_with(":kv") && !is_ident_char(input.as_bytes().get(3).copied()) {
-        return Some((&input[3..], "kv", None));
-    }
-    if let Some(rest) = input.strip_prefix(":p(0)") {
-        return Some((rest, "p0", None));
-    }
-    if let Some(rest) = input.strip_prefix(":p(1)") {
-        return Some((rest, "p", None));
-    }
-    // :p(expr)
-    if input.starts_with(":p(")
-        && let Some((rest, expr)) = try_parse_adverb_expr(&input[2..])
-    {
-        return Some((rest, "p", Some(expr)));
-    }
-    if input.starts_with(":p") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return Some((&input[2..], "p", None));
-    }
-    if let Some(rest) = input.strip_prefix(":k(0)") {
-        return Some((rest, "k0", None));
-    }
-    if let Some(rest) = input.strip_prefix(":k(1)") {
-        return Some((rest, "k", None));
-    }
-    // :k(expr)
-    if input.starts_with(":k(")
-        && let Some((rest, expr)) = try_parse_adverb_expr(&input[2..])
-    {
-        return Some((rest, "k", Some(expr)));
-    }
-    if input.starts_with(":k") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return Some((&input[2..], "k", None));
-    }
-    if let Some(rest) = input.strip_prefix(":v(0)") {
-        return Some((rest, "v0", None));
-    }
-    if let Some(rest) = input.strip_prefix(":v(1)") {
-        return Some((rest, "v", None));
-    }
-    // :v(expr)
-    if input.starts_with(":v(")
-        && let Some((rest, expr)) = try_parse_adverb_expr(&input[2..])
-    {
-        return Some((rest, "v", Some(expr)));
-    }
-    if input.starts_with(":v") && !is_ident_char(input.as_bytes().get(2).copied()) {
-        return Some((&input[2..], "v", None));
-    }
-    None
-}
-
-/// Try to parse a parenthesized expression for a subscript adverb argument.
-/// Input starts at `(`. Returns `(remaining_after_close_paren, expr)`.
-fn try_parse_adverb_expr(input: &str) -> Option<(&str, Expr)> {
-    let r = input.strip_prefix('(')?;
-    let (r, _) = ws(r).ok()?;
-    let (r, expr) = expression(r).ok()?;
-    let (r, _) = ws(r).ok()?;
-    let (r, _) = parse_char(r, ')').ok()?;
-    Some((r, expr))
-}
-
-/// Try to parse an unknown `:identifier` adverb (not k/v/kv/p/exists/delete).
-fn try_parse_unknown_adverb(input: &str) -> Option<(&str, String)> {
-    if !input.starts_with(':') {
-        return None;
-    }
-    let rest = &input[1..];
-    let rest = rest.strip_prefix('!').unwrap_or(rest);
-    if rest.is_empty() || !rest.as_bytes()[0].is_ascii_alphabetic() {
-        return None;
-    }
-    let end = rest
-        .find(|c: char| !c.is_ascii_alphanumeric() && c != '_' && c != '-')
-        .unwrap_or(rest.len());
-    if end == 0 {
-        return None;
-    }
-    let name = &rest[..end];
-    if matches!(name, "k" | "v" | "kv" | "p" | "exists" | "delete") {
-        return None;
-    }
-    let after = &rest[end..];
-    let final_rest = if after.starts_with('(') {
-        after.find(')').map_or(after, |close| &after[close + 1..])
-    } else {
-        after
-    };
-    Some((final_rest, name.to_string()))
-}
-
-/// Determine "element access" vs "slice" from the target and index.
-/// Hash access is always "slice"; array single-element is "element access".
-fn determine_subscript_what(target: &Expr, index_expr: &Expr) -> String {
-    // A *zen* slice (`@a[]` / `%h{}`, modelled as a `Literal(Whatever)` index by
-    // the empty-subscript-with-adverb path) reports "zen slice".
-    if matches!(index_expr, Expr::Literal(Value::Whatever)) {
-        return "zen slice".to_string();
-    }
-    // A whatever slice (`@a[*]`, parsed as a bare Whatever index) reports
-    // "whatever slice". A hash zen slice keeps the bracket-kind descriptor
-    // (`{} slice`), which the runtime downgrades to plain "slice" on a nogo
-    // conflict (see `builtin_subscript_adverb_error`).
-    if matches!(index_expr, Expr::Whatever) {
-        return if matches!(target, Expr::HashVar(_)) {
-            "{} slice".to_string()
-        } else {
-            "whatever slice".to_string()
-        };
-    }
-    if matches!(target, Expr::HashVar(_)) {
-        return "slice".to_string();
-    }
-    // A Range subscript (`@a[1..2]`) is a multi-element slice.
-    if let Expr::Binary { op, .. } = index_expr
-        && matches!(
-            op,
-            TokenKind::DotDot
-                | TokenKind::DotDotCaret
-                | TokenKind::CaretDotDot
-                | TokenKind::CaretDotDotCaret
-        )
-    {
-        return "slice".to_string();
-    }
-    match index_expr {
-        Expr::ArrayLiteral(items) if items.len() != 1 => "slice".to_string(),
-        Expr::ArrayVar(_) => "slice".to_string(),
-        _ => "element access".to_string(),
-    }
-}
-
-/// Normalize an adverb name (strip "not-" prefix and "0" suffix).
-fn normalize_adverb_name(s: &str) -> String {
-    let s = s.strip_prefix("not-").unwrap_or(s);
-    s.strip_suffix('0').unwrap_or(s).to_string()
-}
-
-/// Build a `__mutsu_subscript_adverb_error` call for X::Adverb.
-fn build_adverb_error_call(
-    what: &str,
-    source: &str,
-    nogo: &[String],
-    unexpected: &[String],
-) -> Expr {
-    let mut args = vec![
-        Expr::Literal(Value::str(what.to_string())),
-        Expr::Literal(Value::str(source.to_string())),
-    ];
-    for a in nogo {
-        args.push(Expr::Literal(Value::str(format!("__nogo__{}", a))));
-    }
-    for u in unexpected {
-        args.push(Expr::Literal(Value::str(format!("__unexpected__{}", u))));
-    }
-    Expr::Call {
-        name: Symbol::intern("__mutsu_subscript_adverb_error"),
-        args,
-    }
-}
-
-/// Consume all remaining adverbs (known and unknown) after a subscript.
-fn collect_remaining_adverbs<'a>(
-    start: &'a str,
-    known: &mut Vec<String>,
-    unknown: &mut Vec<String>,
-) -> &'a str {
-    let mut r = start;
-    loop {
-        let r2 = ws(r).map_or(r, |(r2, _)| r2);
-        if let Some((r3, next_adv, _)) = parse_subscript_adverb_with_expr(r2) {
-            known.push(normalize_adverb_name(next_adv));
-            r = r3;
-        } else if let Some((r3, unk_name)) = try_parse_unknown_adverb(r2) {
-            unknown.push(unk_name);
-            r = r3;
-        } else {
-            break;
-        }
-    }
-    r
-}
-
-/// Extract the variable name from a MultiDimIndex target expression.
-fn multidim_target_var_name(target: &Expr) -> String {
-    match target {
-        Expr::ArrayVar(name) => format!("@{}", name),
-        Expr::HashVar(name) => format!("%{}", name),
-        Expr::Var(name) => name.clone(),
-        _ => String::new(),
-    }
-}
-
-enum DeleteAdverb {
-    NoDelete,
-    Delete(Option<Expr>),
-}
-
-fn parse_delete_adverb(input: &str) -> Option<(&str, DeleteAdverb)> {
-    if input.starts_with(":!delete") && !is_ident_char(input.as_bytes().get(8).copied()) {
-        return Some((&input[8..], DeleteAdverb::NoDelete));
-    }
-    if input.starts_with(":delete") && !is_ident_char(input.as_bytes().get(7).copied()) {
-        let mut r = &input[7..];
-        if let Some(r_stripped) = r.strip_prefix('(')
-            && let Ok((r2, _)) = ws(r_stripped)
-            && let Ok((r2, cond)) = expression(r2)
-            && let Ok((r2, _)) = ws(r2)
-            && let Ok((r2, _)) = parse_char(r2, ')')
-        {
-            return Some((r2, DeleteAdverb::Delete(Some(cond))));
-        }
-        // :delete with no argument, or malformed parens (leave for outer parser).
-        if r.starts_with('(') {
-            return None;
-        }
-        r = &input[7..];
-        return Some((r, DeleteAdverb::Delete(None)));
-    }
-    None
-}
-
-fn subscript_adverb_expr_with_cond(expr: Expr, adverb: &'static str, cond: Option<Expr>) -> Expr {
-    // Handle MultiDimIndex: @a[0;0;0]:kv etc.
-    if let Expr::MultiDimIndex { target, dimensions } = expr {
-        let mut args = vec![*target, Expr::Literal(Value::str(adverb.to_string()))];
-        args.extend(dimensions);
-        if let Some(cond_expr) = cond {
-            args.push(Expr::Literal(Value::str("__adverb_cond__".to_string())));
-            args.push(cond_expr);
-        }
-        return Expr::Call {
-            name: Symbol::intern("__mutsu_multidim_subscript_adverb"),
-            args,
-        };
-    }
-    let Expr::Index { target, index, .. } = expr else {
-        return expr;
-    };
-    let var_name = match target.as_ref() {
-        Expr::ArrayVar(name) => Expr::Literal(Value::str(format!("@{}", name))),
-        Expr::HashVar(name) => Expr::Literal(Value::str(format!("%{}", name))),
-        _ => Expr::Literal(Value::Nil),
-    };
-    let mut args = vec![
-        *target,
-        *index,
-        Expr::Literal(Value::str(adverb.to_string())),
-        var_name,
-    ];
-    // When a dynamic condition is provided (e.g., `:k($ok)`), pass it as
-    // a named Pair so the runtime can decide keep_missing at evaluation time.
-    if let Some(cond_expr) = cond {
-        args.push(Expr::Literal(Value::str("__adverb_cond__".to_string())));
-        args.push(cond_expr);
-    }
-    Expr::Call {
-        name: Symbol::intern("__mutsu_subscript_adverb"),
-        args,
-    }
-}
-
-/// Try to parse :exists or :!exists adverb on a subscript expression.
-/// Returns (remaining_input, exists_expr) or None if no adverb found.
-fn try_parse_exists_adverb(input: &str, target: Expr) -> Option<(&str, Expr)> {
-    let r = input;
-    let (r, negated) = if r.starts_with(":!exists") && !is_ident_char(r.as_bytes().get(8).copied())
-    {
-        (&r[8..], true)
-    } else if r.starts_with(":exists") && !is_ident_char(r.as_bytes().get(7).copied()) {
-        (&r[7..], false)
-    } else {
-        return None;
-    };
-    // Check for parameterized argument: :exists(expr)
-    let (r, arg) = if let Some(r_stripped) = r.strip_prefix('(') {
-        let r2 = r_stripped;
-        if let Ok((r2, _)) = ws(r2) {
-            if let Ok((r2, arg_expr)) = expression(r2) {
-                if let Ok((r2, _)) = ws(r2) {
-                    if let Ok((r2, _)) = parse_char(r2, ')') {
-                        (r2, Some(Box::new(arg_expr)))
-                    } else {
-                        (r, None)
-                    }
-                } else {
-                    (r, None)
-                }
-            } else {
-                (r, None)
-            }
-        } else {
-            (r, None)
-        }
-    } else {
-        (r, None)
-    };
-    // Check for secondary adverb
-    let (r, adverb) = parse_exists_secondary_adverb(r);
-    Some((
-        r,
-        Expr::Exists {
-            target: Box::new(target),
-            negated,
-            delete: false,
-            arg,
-            adverb,
-        },
-    ))
-}
-
-fn supports_postfix_call_adverbs(expr: &Expr) -> bool {
-    match expr {
-        Expr::Call { name, .. } => {
-            let n = name.resolve();
-            !n.starts_with("__mutsu_subscript_adverb") && !n.starts_with("__mutsu_multidim")
-        }
-        Expr::MethodCall { .. }
-        | Expr::CallOn { .. }
-        | Expr::HyperMethodCall { .. }
-        | Expr::HyperMethodCallDynamic { .. } => true,
-        _ => false,
-    }
-}
-
-fn auto_invoke_bareword_method_target(expr: Expr) -> Expr {
-    let Expr::BareWord(name) = expr else {
-        return expr;
-    };
-    if crate::parser::stmt::simple::is_user_declared_sub(&name)
-        || crate::parser::stmt::simple::is_imported_function(&name)
-    {
-        return Expr::Call {
-            name: Symbol::intern(&name),
-            args: Vec::new(),
-        };
-    }
-    Expr::BareWord(name)
-}
-
-fn append_call_arg(expr: &mut Expr, arg: Expr) -> bool {
-    match expr {
-        Expr::Call { args, .. } => {
-            args.push(arg);
-            true
-        }
-        Expr::MethodCall { args, .. } => {
-            args.push(arg);
-            true
-        }
-        Expr::CallOn { args, .. } => {
-            args.push(arg);
-            true
-        }
-        Expr::HyperMethodCall { args, .. } => {
-            args.push(arg);
-            true
-        }
-        Expr::HyperMethodCallDynamic { args, .. } => {
-            args.push(arg);
-            true
-        }
-        _ => false,
-    }
-}
-
-/// Result of parsing bracket indices — either a single-dimension index or
-/// a multi-dimensional (semicolon-separated) index.
-enum ParsedBracketIndex {
-    /// Normal single-dimension index (possibly comma-separated list).
-    Single(Expr),
-    /// Multi-dimensional index: dimensions separated by semicolons.
-    /// Each dimension can itself be a comma-separated list.
-    MultiDim(Vec<Expr>),
-}
-
-fn parse_bracket_indices(input: &str) -> PResult<'_, Expr> {
-    let (rest, parsed) = parse_bracket_indices_inner(input)?;
-    match parsed {
-        ParsedBracketIndex::Single(expr) => Ok((rest, expr)),
-        // For callers that don't handle MultiDim, flatten to ArrayLiteral.
-        // The postfix parser will use parse_bracket_indices_inner directly.
-        ParsedBracketIndex::MultiDim(dims) => Ok((rest, Expr::ArrayLiteral(dims))),
-    }
-}
-
-fn parse_bracket_indices_inner(input: &str) -> PResult<'_, ParsedBracketIndex> {
-    // Allow phaser-only blocks (e.g. `%h{ CATCH { } }`) inside subscripts.
-    // In Raku this evaluates the block which returns Nil and then indexes
-    // the hash with Nil; we represent it as a Whatever placeholder to keep
-    // parsing alive (the evaluation will produce a Nil result at runtime).
-    {
-        let probe = input.trim_start();
-        if probe.starts_with("CATCH") || probe.starts_with("CONTROL") {
-            let kw_len = if probe.starts_with("CATCH") { 5 } else { 7 };
-            let after_kw = &probe[kw_len..];
-            if after_kw.is_empty()
-                || after_kw.starts_with(' ')
-                || after_kw.starts_with('\t')
-                || after_kw.starts_with('{')
-            {
-                // Skip whitespace, then parse the phaser body block and discard.
-                let (r, _) = ws(after_kw)?;
-                if r.starts_with('{') {
-                    // Skip a balanced brace block.
-                    let bytes = r.as_bytes();
-                    let mut depth = 0i32;
-                    let mut i = 0;
-                    while i < bytes.len() {
-                        match bytes[i] {
-                            b'{' => depth += 1,
-                            b'}' => {
-                                depth -= 1;
-                                if depth == 0 {
-                                    i += 1;
-                                    break;
-                                }
-                            }
-                            _ => {}
-                        }
-                        i += 1;
-                    }
-                    let after_block = &r[i..];
-                    let (after_block, _) = ws(after_block)?;
-                    return Ok((after_block, ParsedBracketIndex::Single(Expr::Whatever)));
-                }
-            }
-        }
-    }
-    let (r, first) = expression(input)?;
-    let mut current_dim = vec![first];
-    let mut dimensions: Vec<Expr> = Vec::new();
-    let mut has_semicolons = false;
-    let mut has_trailing_comma = false;
-    let mut r = r;
-    loop {
-        let (r2, _) = ws(r)?;
-        if r2.starts_with(',') {
-            let (r3, _) = parse_char(r2, ',')?;
-            let (r3, _) = ws(r3)?;
-            // Handle trailing comma before ']' or ';'
-            if r3.starts_with(']') || r3.starts_with(';') {
-                has_trailing_comma = true;
-                r = r3;
-                continue;
-            }
-            let (r3, next) = expression(r3)?;
-            // `@a[0, 2 ... *]` — the sequence operator's seed is the WHOLE
-            // preceding comma list, not just its immediate left operand
-            // (raku parses this as `(0, 2) ... *`). When a comma item turns out
-            // to be a sequence, fold the items gathered so far into its seed.
-            if let Expr::Binary {
-                left,
-                op: op @ (TokenKind::DotDotDot | TokenKind::DotDotDotCaret),
-                right,
-            } = next
-            {
-                let mut seed = std::mem::take(&mut current_dim);
-                seed.push(*left);
-                current_dim = vec![Expr::Binary {
-                    left: Box::new(Expr::ArrayLiteral(seed)),
-                    op,
-                    right,
-                }];
-            } else {
-                current_dim.push(next);
-            }
-            r = r3;
-            continue;
-        }
-        if r2.starts_with(';') && !r2.starts_with(";;") {
-            has_semicolons = true;
-            // Finish current dimension
-            let dim_expr = if current_dim.len() == 1 {
-                current_dim.remove(0)
-            } else {
-                Expr::ArrayLiteral(std::mem::take(&mut current_dim))
-            };
-            dimensions.push(dim_expr);
-            current_dim = Vec::new();
-            let (r3, _) = parse_char(r2, ';')?;
-            let (r3, _) = ws(r3)?;
-            let (r3, next) = expression(r3)?;
-            current_dim.push(next);
-            r = r3;
-            continue;
-        }
-        if has_semicolons {
-            // Finish last dimension
-            let dim_expr = if current_dim.len() == 1 {
-                current_dim.remove(0)
-            } else {
-                Expr::ArrayLiteral(current_dim)
-            };
-            dimensions.push(dim_expr);
-            return Ok((r2, ParsedBracketIndex::MultiDim(dimensions)));
-        }
-        return Ok((
-            r2,
-            ParsedBracketIndex::Single(if current_dim.len() == 1 && !has_trailing_comma {
-                current_dim.remove(0)
-            } else {
-                Expr::ArrayLiteral(current_dim)
-            }),
-        ));
-    }
-}
-
-/// Result of parsing a quoted method name.
-pub(crate) enum QuotedMethodName {
-    /// Static method name (single-quoted or double-quoted without interpolation)
-    Static(String),
-    /// Dynamic method name (double-quoted with interpolation)
-    Dynamic(Expr),
-}
-
-pub(crate) fn parse_quoted_method_name(input: &str) -> Option<(&str, QuotedMethodName)> {
-    if input.starts_with('\'') {
-        // Single-quoted: no interpolation
-        let start = 1;
-        let mut end = None;
-        for (i, c) in input[start..].char_indices() {
-            if c == '\'' {
-                end = Some(i);
-                break;
-            }
-        }
-        let end = end?;
-        let content = &input[start..start + end];
-        let rest = &input[start + end + 1..];
-        return Some((rest, QuotedMethodName::Static(content.to_string())));
-    }
-    if input.starts_with('"') {
-        // Double-quoted: may have interpolation
-        let start = 1;
-        let mut escaped = false;
-        let mut end = None;
-        for (i, c) in input[start..].char_indices() {
-            if escaped {
-                escaped = false;
-                continue;
-            }
-            if c == '\\' {
-                escaped = true;
-                continue;
-            }
-            if c == '"' {
-                end = Some(i);
-                break;
-            }
-        }
-        let end = end?;
-        let content = &input[start..start + end];
-        let rest = &input[start + end + 1..];
-        // Check if content has interpolation
-        if content.contains('$') || content.contains('@') || content.contains('{') {
-            use super::super::primary::string::interpolate_string_content_with_modes;
-            let name_expr = interpolate_string_content_with_modes(content, true, true);
-            return Some((rest, QuotedMethodName::Dynamic(name_expr)));
-        }
-        return Some((rest, QuotedMethodName::Static(content.to_string())));
-    }
-    None
-}
-
-fn parse_private_method_name(input: &str) -> Option<(&str, String)> {
-    let mut rest = input;
-    let mut name = String::new();
-    let mut first = true;
-    loop {
-        let (r, part) =
-            take_while1(rest, |c: char| c.is_alphanumeric() || c == '_' || c == '-').ok()?;
-        if !first {
-            name.push_str("::");
-        }
-        name.push_str(part);
-        first = false;
-        rest = r;
-        if let Some(r2) = rest.strip_prefix("::") {
-            rest = r2;
-            continue;
-        }
-        break;
-    }
-    Some((rest, name))
-}
-
-/// Parse the operator name from a prefix-as-postfix construct like `<->`, `«~»`, `<<~>>`, `["~"]`.
-/// Input starts after the `:` in `.:<->`.
-fn parse_prefix_as_postfix(input: &str) -> Option<(&str, String)> {
-    // <<op>> (must be checked before <op>)
-    if let Some(rest) = input.strip_prefix("<<") {
-        // Handle both <<op>> and <<'op'>>
-        let end = rest.find(">>")?;
-        let mut op = &rest[..end];
-        // Strip quotes if present: <<'op'>>
-        if let Some(inner) = op.strip_prefix('\'').and_then(|s| s.strip_suffix('\'')) {
-            op = inner;
-        }
-        if op.is_empty() {
-            return None;
-        }
-        return Some((&rest[end + 2..], op.to_string()));
-    }
-    // <op>
-    if let Some(rest) = input.strip_prefix('<') {
-        let end = rest.find('>')?;
-        let op = &rest[..end];
-        if op.is_empty() {
-            return None;
-        }
-        return Some((&rest[end + 1..], op.to_string()));
-    }
-    // «op» (U+00AB / U+00BB)
-    if let Some(rest) = input.strip_prefix('\u{00AB}') {
-        let end = rest.find('\u{00BB}')?;
-        let op = &rest[..end];
-        if op.is_empty() {
-            return None;
-        }
-        return Some((&rest[end + '\u{00BB}'.len_utf8()..], op.to_string()));
-    }
-    // ["op"]
-    if let Some(rest) = input.strip_prefix('[')
-        && let Some(rest) = rest.strip_prefix('"')
-    {
-        let end = rest.find('"')?;
-        let op = &rest[..end];
-        let rest = &rest[end + 1..];
-        let rest = rest.strip_prefix(']')?;
-        if op.is_empty() {
-            return None;
-        }
-        return Some((rest, op.to_string()));
-    }
-    None
-}
-
-fn is_postfix_operator_char(c: char) -> bool {
-    if c.is_whitespace() || c.is_alphanumeric() || c == '_' {
-        return false;
-    }
-    // Exclude characters that are infix operators or special symbols
-    // U+2026 HORIZONTAL ELLIPSIS (…) is the sequence infix operator
-    // U+221E INFINITY (∞) is a numeric literal, not an operator char
-    if c == '\u{2026}' || c == '\u{221E}' {
-        return false;
-    }
-    !matches!(
-        c,
-        '.' | ','
-            | ';'
-            | ':'
-            | '('
-            | ')'
-            | '['
-            | ']'
-            | '{'
-            | '}'
-            | '"'
-            | '\''
-            | '\\'
-            | '$'
-            | '@'
-            | '%'
-            | '&'
-            | '#'
-    )
-}
-
-fn is_postfix_operator_boundary(rest: &str) -> bool {
-    rest.is_empty()
-        || rest.starts_with(|c: char| {
-            c.is_whitespace() || matches!(c, '.' | ')' | '}' | ']' | ',' | ';' | ':')
-        })
-}
-
-fn has_ternary_else_after(input: &str) -> bool {
-    let mut rest = input;
-    while let Ok((next, _)) = ws(rest) {
-        if next.len() == rest.len() {
-            break;
-        }
-        rest = next;
-    }
-    rest.starts_with("!!")
-}
-
-fn parse_custom_postfix_operator(input: &str) -> Option<(String, usize)> {
-    // Don't consume characters that are closing delimiters of circumfix operators
-    if crate::parser::stmt::simple::is_circumfix_close_delimiter(input) {
-        return None;
-    }
-
-    if input.starts_with('!') {
-        return Some(("!".to_string(), '!'.len_utf8()));
-    }
-
-    let mut chars = input.chars();
-    let first = chars.next()?;
-    if first.is_ascii() || !is_postfix_operator_char(first) {
-        return None;
-    }
-
-    let mut len = first.len_utf8();
-    for c in chars {
-        if c.is_ascii() || !is_postfix_operator_char(c) {
-            break;
-        }
-        len += c.len_utf8();
-    }
-    Some((input[..len].to_string(), len))
-}
-
-fn atomic_var_name(expr: &Expr) -> Option<String> {
-    match expr {
-        Expr::Var(name) => Some(name.clone()),
-        _ => None,
-    }
-}
-
-/// Wrap a `.=` method call result in the appropriate assignment expression.
-/// For simple variables, generates `AssignExpr { name, expr }`.
-/// For index expressions, generates an `IndexAssign` wrapped in a `DoBlock`.
-/// For non-lvalue targets, returns the method call as-is.
-fn wrap_dot_assign(target: Expr, method_call_fn: impl FnOnce(Expr) -> Expr) -> Expr {
-    // Unwrap Grouped to get at the inner expression
-    let target = match target {
-        Expr::Grouped(inner) => *inner,
-        other => other,
-    };
-    match &target {
-        Expr::Var(name) => Expr::AssignExpr {
-            name: name.clone(),
-            expr: Box::new(method_call_fn(target)),
-            is_bind: false,
-        },
-        Expr::ArrayVar(name) => Expr::AssignExpr {
-            name: format!("@{}", name),
-            expr: Box::new(method_call_fn(target)),
-            is_bind: false,
-        },
-        Expr::HashVar(name) => Expr::AssignExpr {
-            name: format!("%{}", name),
-            expr: Box::new(method_call_fn(target)),
-            is_bind: false,
-        },
-        Expr::Index {
-            target: idx_target,
-            index,
-            is_positional,
-        } => {
-            use std::sync::atomic::Ordering;
-            let tmp_idx = format!(
-                "__mutsu_idx_{}",
-                super::super::stmt::simple::TMP_INDEX_COUNTER.fetch_add(1, Ordering::Relaxed)
-            );
-            let tmp_idx_expr = Expr::Var(tmp_idx.clone());
-            let lhs_expr = Expr::Index {
-                target: idx_target.clone(),
-                index: Box::new(tmp_idx_expr.clone()),
-                is_positional: *is_positional,
-            };
-            let assigned_value = method_call_fn(lhs_expr);
-            Expr::DoBlock {
-                body: vec![
-                    Stmt::VarDecl {
-                        name: tmp_idx.clone(),
-                        expr: *index.clone(),
-                        type_constraint: None,
-                        is_state: false,
-                        is_our: false,
-                        is_dynamic: false,
-                        is_export: false,
-                        export_tags: Vec::new(),
-                        custom_traits: Vec::new(),
-                        where_constraint: None,
-                    },
-                    Stmt::Expr(Expr::IndexAssign {
-                        target: idx_target.clone(),
-                        index: Box::new(tmp_idx_expr),
-                        value: Box::new(assigned_value),
-                        is_positional: *is_positional,
-                    }),
-                ],
-                label: None,
-            }
-        }
-        // ($var = expr).=method => evaluate the assignment, then $var = $var.method
-        Expr::AssignExpr { name, .. } => {
-            let assign_name = name.clone();
-            let var_expr = if let Some(rest) = assign_name.strip_prefix('@') {
-                Expr::ArrayVar(rest.to_string())
-            } else if let Some(rest) = assign_name.strip_prefix('%') {
-                Expr::HashVar(rest.to_string())
-            } else {
-                Expr::Var(assign_name.clone())
-            };
-            let method_result = method_call_fn(var_expr);
-            Expr::DoBlock {
-                body: vec![
-                    Stmt::Expr(target),
-                    Stmt::Expr(Expr::AssignExpr {
-                        name: assign_name,
-                        expr: Box::new(method_result),
-                        is_bind: false,
-                    }),
-                ],
-                label: None,
-            }
-        }
-        // A parenthesized list of lvalues (`($x, $y) .= reverse`) applies the
-        // method to the WHOLE list and assigns the result back element-wise
-        // (`($x, $y) = ($x, $y).reverse`), not per-element.
-        Expr::ArrayLiteral(_) => {
-            let lhs = target.clone();
-            let value = method_call_fn(target);
-            Expr::Call {
-                name: Symbol::intern("__mutsu_assign_callable_lvalue"),
-                args: vec![lhs, Expr::ArrayLiteral(Vec::new()), value],
-            }
-        }
-        _ => method_call_fn(target),
-    }
-}
-
-/// Parse `.=` mutating method call after the `=` has been identified.
-/// `input` starts after the `=` character. `expr` is the target expression.
-/// Kept as a separate function to avoid bloating the postfix loop's stack frame.
-#[inline(never)]
-fn parse_dot_assign<'a>(input: &'a str, expr: Expr) -> PResult<'a, Expr> {
-    let (r, _) = ws(input)?;
-    // Parse quoted method name
-    if let Some((r_after, qname)) = parse_quoted_method_name(r) {
-        let (r_after, _) = ws(r_after)?;
-        let (r_final, args) = if r_after.starts_with('(') {
-            let (r2, _) = parse_char(r_after, '(')?;
-            let (r2, _) = ws(r2)?;
-            let (r2, a) = parse_call_arg_list(r2)?;
-            let (r2, _) = ws(r2)?;
-            let (r2, _) = parse_char(r2, ')')?;
-            (r2, a)
-        } else {
-            (r_after, vec![])
-        };
-        let result = match qname {
-            QuotedMethodName::Static(mname) => {
-                let mname_sym = Symbol::intern(&mname);
-                wrap_dot_assign(expr, |target| Expr::MethodCall {
-                    target: Box::new(target),
-                    name: mname_sym,
-                    args: args.clone(),
-                    modifier: None,
-                    quoted: true,
-                })
-            }
-            QuotedMethodName::Dynamic(name_expr) => {
-                wrap_dot_assign(expr, |target| Expr::DynamicMethodCall {
-                    target: Box::new(target),
-                    name_expr: Box::new(name_expr.clone()),
-                    args: args.clone(),
-                    modifier: None,
-                })
-            }
-        };
-        return Ok((r_final, result));
-    }
-    // `.= &sub` / `.= $meth` — mutating method-call-via-sub or dynamic method
-    // name (`$x .= &f` means `$x = $x.&f`). The name is a sigiled expression.
-    if r.starts_with('&') || r.starts_with('$') || r.starts_with('@') || r.starts_with('%') {
-        let (r_name, name_expr) = crate::parser::primary::primary(r)?;
-        let (r_after, _) = ws(r_name)?;
-        let (r_final, args) = if r_after.starts_with('(') {
-            let (r2, _) = parse_char(r_after, '(')?;
-            let (r2, _) = ws(r2)?;
-            let (r2, a) = parse_call_arg_list(r2)?;
-            let (r2, _) = ws(r2)?;
-            let (r2, _) = parse_char(r2, ')')?;
-            (r2, a)
-        } else {
-            (r_name, vec![])
-        };
-        let result = wrap_dot_assign(expr, |target| Expr::DynamicMethodCall {
-            target: Box::new(target),
-            name_expr: Box::new(name_expr.clone()),
-            args: args.clone(),
-            modifier: None,
-        });
-        return Ok((r_final, result));
-    }
-    // Parse regular method name
-    let (r, method_name) = crate::parser::primary::var::parse_ident_with_hyphens(r)?;
-    let r_before_ws = r;
-    let (r, _) = ws(r)?;
-    let (r_final, args) = if r.starts_with('(') {
-        let (r2, _) = parse_char(r, '(')?;
-        let (r2, _) = ws(r2)?;
-        let (r2, a) = parse_call_arg_list(r2)?;
-        let (r2, _) = ws(r2)?;
-        let (r2, _) = parse_char(r2, ')')?;
-        (r2, a)
-    } else if r_before_ws.starts_with(':') && !r_before_ws.starts_with("::") {
-        // Colon-arg syntax: .=method: arg (no space before colon)
-        let r2 = &r[1..];
-        let (r2, _) = ws(r2)?;
-        let (r2, first_arg) = expression(r2)?;
-        let mut args = vec![first_arg];
-        let mut r_inner = r2;
-        loop {
-            let (r3, _) = ws(r_inner)?;
-            if r3.starts_with(':')
-                && !r3.starts_with("::")
-                && let Ok((r4, arg)) = crate::parser::primary::misc::colonpair_expr(r3)
-            {
-                args.push(arg);
-                r_inner = r4;
-                continue;
-            }
-            if !r3.starts_with(',') {
-                r_inner = r3;
-                break;
-            }
-            let r3 = &r3[1..];
-            let (r3, _) = ws(r3)?;
-            if r3.starts_with(';') || r3.starts_with('}') || r3.is_empty() {
-                r_inner = r3;
-                break;
-            }
-            let (r3, next) = expression(r3)?;
-            args.push(next);
-            r_inner = r3;
-        }
-        (r_inner, args)
-    } else if r.starts_with(':') && !r.starts_with("::") {
-        // Fake-infix adverb form (space before ':'): .=method :key<val>
-        let mut args = Vec::new();
-        let mut r_inner = r;
-        while r_inner.starts_with(':') && !r_inner.starts_with("::") {
-            if let Ok((r2, arg)) = crate::parser::primary::misc::colonpair_expr(r_inner) {
-                args.push(arg);
-                let (r3, _) = ws(r2)?;
-                r_inner = r3;
-            } else {
-                break;
-            }
-        }
-        (r_inner, args)
-    } else {
-        (r, vec![])
-    };
-    let method_sym = Symbol::intern(method_name);
-    let result = wrap_dot_assign(expr, |target| Expr::MethodCall {
-        target: Box::new(target),
-        name: method_sym,
-        args: args.clone(),
-        modifier: None,
-        quoted: false,
-    });
-    Ok((r_final, result))
-}
-
-pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
+pub(in crate::parser::expr) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
     if let Some(rest) = input.strip_prefix("++⚛") {
         let (rest, expr) = postfix_expr(rest)?;
         if let Some(name) = atomic_var_name(&expr) {
@@ -1276,15 +125,15 @@ pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
             let (mut rest, arg) = match prec_level {
                 Some(level) if level <= crate::parser::stmt::simple::PREC_ADDITIVE => {
                     // Looser than additive: grab everything up to additive level
-                    super::precedence::loose_prefix_operand(rest, level)?
+                    crate::parser::expr::precedence::loose_prefix_operand(rest, level)?
                 }
                 Some(level) if level <= crate::parser::stmt::simple::PREC_MULTIPLICATIVE => {
                     // Between additive and multiplicative
-                    super::precedence::multiplicative_operand(rest)?
+                    crate::parser::expr::precedence::multiplicative_operand(rest)?
                 }
                 Some(level) if level <= crate::parser::stmt::simple::PREC_POWER => {
                     // Between multiplicative and power
-                    super::precedence::power_operand(rest)?
+                    crate::parser::expr::precedence::power_operand(rest)?
                 }
                 _ => prefix_expr(rest)?,
             };
@@ -1420,7 +269,7 @@ pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
         let r = &input[4..];
         let (r, _) = ws(r)?;
         // Try to parse as `lazy for ...` first
-        if let Ok((r2, stmt)) = super::super::stmt::lazy_for_stmt_pub(r) {
+        if let Ok((r2, stmt)) = crate::parser::stmt::lazy_for_stmt_pub(r) {
             return Ok((r2, Expr::DoStmt(Box::new(stmt))));
         }
         let (r, expr) = expression_no_sequence(r)?;
@@ -1465,8 +314,8 @@ pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
     if input.starts_with('^') && !input.starts_with("^..") {
         // Allow optional whitespace after prefix ^: both `^10` and `^ 10` are valid.
         let rest = input[1..].trim_start_matches([' ', '\t']);
-        let parsed_operand =
-            super::precedence_meta_ops::power_expr_tight(rest).or_else(|_| prefix_expr(rest));
+        let parsed_operand = crate::parser::expr::precedence_meta_ops::power_expr_tight(rest)
+            .or_else(|_| prefix_expr(rest));
         if let Ok((rest, expr)) = parsed_operand {
             let mut node = Expr::Binary {
                 left: Box::new(Expr::Literal(Value::Int(0))),
@@ -1474,8 +323,10 @@ pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
                 right: Box::new(expr.clone()),
             };
             // ^* is WhateverCode: given N, returns ^N (unlike 0..^* which is a Range)
-            if super::contains_whatever(&expr) || super::is_whatever(&expr) {
-                node = super::wrap_whatevercode(&node);
+            if crate::parser::expr::contains_whatever(&expr)
+                || crate::parser::expr::is_whatever(&expr)
+            {
+                node = crate::parser::expr::wrap_whatevercode(&node);
             }
             return postfix_expr_continue(rest, node);
         }
@@ -1505,7 +356,7 @@ pub(super) fn prefix_expr(input: &str) -> PResult<'_, Expr> {
             },
         ));
     }
-    super::precedence_meta_ops::power_expr(input)
+    crate::parser::expr::precedence_meta_ops::power_expr(input)
 }
 
 /// Like `postfix_expr` but does NOT allow whitespace-separated dotty method
@@ -1516,7 +367,7 @@ fn postfix_expr_tight(input: &str) -> PResult<'_, Expr> {
 }
 
 /// Public wrapper for `postfix_expr_tight`, used by `power_expr_tight`.
-pub(super) fn postfix_expr_tight_pub(input: &str) -> PResult<'_, Expr> {
+pub(in crate::parser::expr) fn postfix_expr_tight_pub(input: &str) -> PResult<'_, Expr> {
     postfix_expr_tight(input)
 }
 
@@ -2129,7 +980,7 @@ fn postfix_expr_loop(mut rest: &str, mut expr: Expr, allow_ws_dot: bool) -> PRes
                     rest = r_name;
                     continue;
                 }
-                let (r_name, name_expr) = super::super::primary::primary(r)?;
+                let (r_name, name_expr) = crate::parser::primary::primary(r)?;
                 let (r_name, _) = ws(r_name)?;
                 if r_name.starts_with('(') {
                     let (r_name, _) = parse_char(r_name, '(')?;
@@ -3253,7 +2104,7 @@ fn postfix_expr_loop(mut rest: &str, mut expr: Expr, allow_ws_dot: bool) -> PRes
             // Only match when NOT preceded by dot (dotted form is a method call)
             if !after_hyper.starts_with('.')
                 && let Some((full_name, consumed)) =
-                    super::super::stmt::simple::match_user_declared_postfix_op(after_hyper)
+                    crate::parser::stmt::simple::match_user_declared_postfix_op(after_hyper)
             {
                 expr = Expr::HyperMethodCall {
                     target: Box::new(expr),
@@ -3392,7 +2243,7 @@ fn postfix_expr_loop(mut rest: &str, mut expr: Expr, allow_ws_dot: bool) -> PRes
             // `@a».???` calls `postfix:<???>`, but `@a».foo` looks up a method.
             if modifier.is_none()
                 && let Some((full_name, consumed)) =
-                    super::super::stmt::simple::match_user_declared_postfix_op(r)
+                    crate::parser::stmt::simple::match_user_declared_postfix_op(r)
             {
                 let inner = full_name
                     .strip_prefix("postfix:<")
@@ -3563,7 +2414,7 @@ fn postfix_expr_loop(mut rest: &str, mut expr: Expr, allow_ws_dot: bool) -> PRes
                     rest = r_name;
                     continue;
                 }
-                let (r_name, name_expr) = super::super::primary::primary(r)?;
+                let (r_name, name_expr) = crate::parser::primary::primary(r)?;
                 let (r_name, _) = ws(r_name)?;
                 if r_name.starts_with('(') {
                     let (r_name, _) = parse_char(r_name, '(')?;
